IRV: Fix detection of round loser
The results sorting was removing the array keys, which contained the
candidates, removing the ability to ever do a second round of counting.

Because the results include all candidates, the bottom candidates may
note have any ballots cast in their favour. That would prematurely end
the loop and stop the count. Make sure to select the candidate with the
lowest non-zero amount of ballots instead.